Ajustes en test unitarios campaign

+++ src/main/java/com/jayway/hopelink/campaign/contract/document/CampaignDocument.java
@@ -5,6 +5,8 @@ import com.jayway.hopelink.campaign.enums.RecipientType;
 import lombok.*;
 import org.springframework.data.annotation.Id;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
@@ -14,7 +16,7 @@ public class CampaignDocument {
 
     @Id
     private String campaignId;
-    private Double targetAmount;
+    private BigDecimal targetAmount;
     private Integer postalCode;
     private Category category;
     private RecipientType recipientType;
